Alpha clipped tree prop (#449)
* sandbox: Add tree model

* rend: Support alpha clip in the geometry pass

* sandbox: Add tree prefab

* sandbox: Spawn tree props

* rend: Remove unused pragma from depth_only frag shader

* rend: Allow relaxed shader interfaces

Meaning the vertex shader is allowed to produce more data then
the fragment shader consumes.

* rend: Produce texcoords in position only shaders

Slightly violates the name (as its no longer fully position-only) so
needs some cleanup in the future but it allows alpha-clip shaders
in the shadow pass.

* rend: Add a depth-only shader with alpha clipping

* rend: Detect shaders that are using discard

* rend: Detect graphics that may discard fragments

* rend: Fix invalid binding in depth_only_clip.frag

* rend: Add clipped shadow graphic

* rend: Minor cleanup in build-in resources

* rend: Rename position shaders to 'dyn_depth'

* rend: Cleanup in shadow graphic names

* rend: Dynamic image shader binding support

* rend: Use dynamic image binding in dyn_depth_clip.frag

* rend: Store shader specialization binding using 8 bit

* rend: Remove discard instructions from ambient shader

* asset: Track kill instruction id

* asset: Parse spv labels

* asset: Track declaration instruction for spv id's

* asset: Detect backwards branches in spv shaders

* asset: Detect backwards branch in switch default label

* asset: Detect backwards branches in switches

Not likely but allowed by the spec

* asset: Track spirv spec constant branches

* asset: Track spec const mask for kill instructions

* rend: Add rvk_shader_may_kill util

* asset: Track default val for boolean spec constants

* rend: Check if kill instruction is unreachable given spec consts

* rend: Exclude alpha clip geo from shadow pass

* rend: Add todo about merging dynamic and draw sets

* rend: Texture prepare support

* rend: Add dynamic image support to pass

* rend: Support alpha clipped shadow draw

* sandbox: Enable backface culling on the tree's

+++ libs/asset/include/asset_graphic.h
@@ -75,7 +75,7 @@ typedef enum {
 
 typedef struct {
   String name;
-  u32    binding;
+  u8     binding;
   f64    value;
 } AssetGraphicOverride;
